Feat: Add icon to keep track of question evaluation
Icon in Header component that is updated when a user marks a question as correct, incorrect, or if the user answers a question/input answer correctly it automatically updates.  (still need to implement Multiple Choice).  If a question is not yet evaluated, or the user skips evaluation, the icon is empty